+++ app/components/Views/confirmations/hooks/transactions/useTransactionMetadataRequest.ts
@@ -22,3 +22,13 @@ export function useTransactionMetadataRequest() {
 
   return transactionMetadata as TransactionMeta;
 }
+
+export function useTransactionMetadataOrThrow() {
+  const transactionMetadata = useTransactionMetadataRequest();
+
+  if (!transactionMetadata) {
+    throw new Error('Transaction approval request not found');
+  }
+
+  return transactionMetadata;
+}
